Move exceptions handlers
