+++ src/server.ts
@@ -2,8 +2,9 @@ import path from "path"
 import Fastify from "fastify"
 import autoLoad from "@fastify/autoload"
 import { serializerCompiler, validatorCompiler, ZodTypeProvider } from "fastify-type-provider-zod"
-import statuses from "statuses"
-import { ZodError, ZodIssueCode } from "zod"
+
+import errorHandler from "./exceptions/errorHandler"
+import notFoundHandler from "./exceptions/notFoundHandler"
 
 import statusRoutes from "./modules/status/status.route"
 import routes from "./modules/routes"
@@ -35,44 +36,10 @@ function buildServer() {
   })
 
   // Set custom error handler
-  server.setErrorHandler((error, _request, reply) => {
-    interface ParsedError {
-      code: ZodIssueCode
-      message: string
-      path: Array<string | number>
-    }
-
-    const statusCode = error?.statusCode || 500
-    let errors = [error] || "Something went wrong"
-
-    if (error instanceof ZodError) {
-      const parsedErrors = JSON.parse(error?.message).map((err: ParsedError) => ({
-        code: err.code,
-        message: err.message,
-        path: err.path
-      }))
-
-      errors = parsedErrors
-    }
-
-    reply.code(statusCode).send({
-      errors,
-      status: statuses(statusCode) || "Unknown error",
-      statusCode
-    })
-  })
+  server.setErrorHandler(errorHandler)
 
   // Set custom not found handler to match our error format
-  server.setNotFoundHandler((request, reply) => {
-    const { url, method } = request.raw
-    const statusCode = 404
-
-    reply.code(statusCode).send({
-      errors: [{ message: `Route ${method}:${url} not found` }],
-      status: statuses(statusCode),
-      statusCode
-    })
-  })
+  server.setNotFoundHandler(notFoundHandler)
 
   // Register status route
   server.register(statusRoutes, { prefix: "status" })
